changes till make s2

+++ src/utils/eda_s2_utils.py
@@ -347,17 +347,21 @@ def download_nhd_shape(area_list, content_list):
             f.write(res.content)
         print(f'download NHD for {i} - {name}')
 
-        # extract the flowline-related files 
-        with zipfile.ZipFile(file_path, 'r') as z:
-            contents = z.namelist()
-            
-            for item in content_list:
-                if item in contents:
-                    z.extract(item, dir)
-                    print(f'\nextract {i} {item}')
+        # extract the flowline-related files
+        if zipfile.is_zipfile(file_path):
+            print(f'File {name} is a valid Zip')
+            with zipfile.ZipFile(file_path, 'r') as z:
+                contents = z.namelist()
+                
+                for item in content_list:
+                    if item in contents:
+                        z.extract(item, dir)
+                        print(f'\nextract {i} {item}')
 
-        # delete the ZIP file
-        os.remove(file_path)
+            # delete the ZIP file
+            os.remove(file_path)
+        else:
+            print("Not a zip file")
 
     # explore the contents within the ZIP file
     print(f'list all contents within the ZIP file for {i}')
